Cleanup
- moved logic handling crs to filters.py
- removed some doctrings, added others
- removed unnecessary variables

leftovers removed
