+++ src/django_oapif/decorators.py
@@ -1,11 +1,6 @@
-import re
-from os import getenv
 from typing import Any, Callable, Dict, Optional
 
-from django.contrib.gis.geos import Polygon
 from django.db.models import Model
-from django.http import HttpResponseBadRequest
-from pyproj import CRS, Transformer
 from rest_framework import viewsets
 from rest_framework_gis.serializers import GeoFeatureModelSerializer
 
@@ -13,59 +8,6 @@ from django_oapif.mixins import OAPIFDescribeModelViewSetMixin
 from django_oapif.urls import oapif_router
 
 from .filters import BboxFilterBackend
-
-# taken from https://github.com/geopython/pygeoapi/blob/953b6fa74d2ce292d8f566c4f4d3bcb4161d6e95/pygeoapi/util.py#L90
-CRS_AUTHORITY = [
-    "AUTO",
-    "EPSG",
-    "OGC",
-]
-
-CRS_URI_PATTERN = re.compile(
-    (
-        rf"^http://www.opengis\.net/def/crs/"
-        rf"(?P<auth>{'|'.join(CRS_AUTHORITY)})/"
-        rf"[\d|\.]+?/(?P<code>\w+?)$"
-    )
-)
-
-
-# taken from
-def get_crs_from_uri(uri: str) -> CRS:
-    """
-    Get a `pyproj.CRS` instance from a CRS URI.
-    Author: @MTachon
-
-    :param uri: Uniform resource identifier of the coordinate
-        reference system.
-    :type uri: str
-
-    :raises `CRSError`: Error raised if no CRS could be identified from the
-        URI.
-
-    :returns: `pyproj.CRS` instance matching the input URI.
-    :rtype: `pyproj.CRS`
-    """
-
-    try:
-        crs = CRS.from_authority(*CRS_URI_PATTERN.search(uri).groups())
-    except RuntimeError:
-        msg = (
-            f"CRS could not be identified from URI {uri!r} "
-            f"(Authority: {CRS_URI_PATTERN.search(uri).group('auth')!r}, "
-            f"Code: {CRS_URI_PATTERN.search(uri).group('code')!r})."
-        )
-        raise RuntimeError(msg)
-    except AttributeError:
-        msg = (
-            f"CRS could not be identified from URI {uri!r}. CRS URIs must "
-            "follow the format "
-            "'http://www.opengis.net/def/crs/{authority}/{version}/{code}' "
-            "(see https://docs.opengeospatial.org/is/18-058r1/18-058r1.html#crs-overview)."  # noqa
-        )
-        raise AttributeError(msg)
-    else:
-        return crs
 
 
 def register_oapif_viewset(
@@ -102,16 +44,17 @@ def register_oapif_viewset(
                 fields = "__all__"
                 geo_field = "geom"
 
-        """ ON HOLD, WAITING ON GeoFeatureModelSerializer to admit of null geometries """
-        # class AutoNoGeomSerializer(ModelSerializer):
-        #     class Meta:
-        #         model = Model
-        #         fields = "__all__"
-
-        # if skip_geom:
-        #     viewset_serializer_class = AutoNoGeomSerializer
-        #     viewset_oapif_geom_lookup = None
-        # else:
+        # ON HOLD, WAITING ON GeoFeatureModelSerializer to admit of null geometries
+        """
+        class AutoNoGeomSerializer(ModelSerializer):
+            class Meta:
+                model = Model
+                fields = "__all__
+        if skip_geom:
+            viewset_serializer_class = AutoNoGeomSerializer
+            viewset_oapif_geom_lookup = None
+        else:
+        """
         viewset_serializer_class = AutoSerializer
         viewset_oapif_geom_lookup = (
             "geom"  # one day this will be retrieved automatically from the serializer
@@ -133,41 +76,13 @@ def register_oapif_viewset(
             # Allowing '.' and '-' in urls
             lookup_value_regex = r"[\w.-]+"
 
-            def get_queryset(self):
-                # Override get_queryset to catch bbox-crs
-                queryset = super().get_queryset()
-
-                if self.request.GET.get("bbox"):
-                    coords = self.request.GET["bbox"].split(",")
-                    user_crs = self.request.GET.get("bbox-crs")
-
-                    if user_crs:
-                        try:
-                            user_crs = get_crs_from_uri(user_crs)
-                        except:
-                            return HttpResponseBadRequest(
-                                "This API supports only EPSG-specified CRS. Make sure to use the appropriate value for the `bbox-crs`query parameter."
-                            )
-                        api_crs = CRS.from_epsg(int(getenv("GEOMETRY_SRID", "2056")))
-                        transformer = Transformer.from_crs(user_crs, api_crs)
-                        LL = transformer.transform(coords[0], coords[1])
-                        UR = transformer.transform(coords[2], coords[3])
-                        my_bbox_polygon = Polygon.from_bbox(
-                            [LL[0], LL[1], UR[0], UR[1]]
-                        )
-
-                    else:
-                        my_bbox_polygon = Polygon.from_bbox(coords)
-
-                    return queryset.filter(geom__intersects=my_bbox_polygon)
-
-                return queryset.all()
-
+        # ON HOLD, WAITING ON GeoFeatureModelSerializer to admit of null geometries
+        """
         # Apply custom serializer attributes
-        # if viewset_serializer_class.__name__ == "AutoNoGeomSerializer":
-        #     for k, v in custom_serializer_attrs.items():
-        #         setattr(AutoNoGeomSerializer.Meta, k, v)
-
+        if viewset_serializer_class.__name__ == "AutoNoGeomSerializer":
+             for k, v in custom_serializer_attrs.items():
+                 setattr(AutoNoGeomSerializer.Meta, k, v)
+        """
         # Apply custom serializer attributes
         for k, v in custom_serializer_attrs.items():
             setattr(AutoSerializer.Meta, k, v)
