+++ src/django_oapif/filters.py
@@ -1,5 +1,10 @@
+from os import getenv
+
 from django.contrib.gis.geos import Polygon
+from pyproj import CRS, Transformer
 from rest_framework.filters import BaseFilterBackend
+
+from .crs_utils import get_crs_from_uri
 
 
 # Adapted from rest_framework_gis.filters.InBBoxFilter
@@ -9,10 +14,20 @@ class BboxFilterBackend(BaseFilterBackend):
         if not bbox_string:
             return queryset
 
-        p1x, p1y, p2x, p2y = (float(n) for n in bbox_string.split(","))
-        # TODO: take into account bbox-crs which may differ from the geom
-        bbox = Polygon.from_bbox((p1x, p1y, p2x, p2y))
-        # TODO: geom shouldn't be hardcoded here
+        coords = tuple(float(n) for n in bbox_string.split(","))
+        user_crs = request.query_params.get("bbox-crs")
+
+        if user_crs:
+            user_crs = get_crs_from_uri(user_crs)
+            api_crs = CRS.from_epsg(int(getenv("GEOMETRY_SRID", "2056")))
+            transformer = Transformer.from_crs(user_crs, api_crs)
+            LL = transformer.transform(coords[0], coords[1])
+            UR = transformer.transform(coords[2], coords[3])
+            bbox = Polygon.from_bbox([LL[0], LL[1], UR[0], UR[1]])
+
+        else:
+            bbox = Polygon.from_bbox(coords)
+
         return queryset.filter(geom__bboverlaps=bbox)
 
     def get_schema_operation_parameters(self, view):
